refactor: move reconstruct command to internal/cli (#529 slice 6)
Lift-and-shift `reconstruct` (command var + init + runReconstruct +
full-table mode) and its four test files from cmd/bintrail (package main)
into internal/cli (package cli), registering it via AddReadCommands so both
the bintrail and the upcoming bintrail-pg binaries share one definition.

The only shared helper this surfaced was parseByteSize (used by both
reconstruct's --chunk-size and agent's --buffer-max-bytes). Extracted it to
cliutil.ParseByteSize (exported) so the cli package and the still-in-main
agent command both reference one copy; removed the agent.go duplicate and
its now-unused math/strconv imports.

Behavior unchanged: same flags, same validation, same output. The
read_commands_test guard now pins reconstruct onto the real rootCmd.

Part of #529.
