+++ cmd/bintrail/agent_test.go
@@ -674,38 +674,3 @@ func TestAgentServerUUIDHelpWarnsAboutSilentIgnore(t *testing.T) {
 		}
 	}
 }
-
-func TestParseByteSize(t *testing.T) {
-	tests := []struct {
-		input string
-		want  int64
-	}{
-		{"0", 0},
-		{"", 0},
-		{"1024", 1024},
-		{"256MB", 256 << 20},
-		{"256mb", 256 << 20},
-		{"1GB", 1 << 30},
-		{"1gb", 1 << 30},
-		{"512KB", 512 << 10},
-	}
-	for _, tt := range tests {
-		got, err := parseByteSize(tt.input)
-		if err != nil {
-			t.Errorf("parseByteSize(%q): unexpected error: %v", tt.input, err)
-			continue
-		}
-		if got != tt.want {
-			t.Errorf("parseByteSize(%q) = %d, want %d", tt.input, got, tt.want)
-		}
-	}
-}
-
-func TestParseByteSize_invalid(t *testing.T) {
-	for _, input := range []string{"abc", "-1GB", "not_a_number", "1.5GB", "9999999999GB"} {
-		_, err := parseByteSize(input)
-		if err == nil {
-			t.Errorf("parseByteSize(%q): expected error", input)
-		}
-	}
-}
