+++ internal/cli/reconstruct.go
@@ -1,4 +1,4 @@
-package main
+package cli
 
 import (
 	"errors"
@@ -13,7 +13,6 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/dbtrail/dbtrail/internal/baseline"
-	"github.com/dbtrail/dbtrail/internal/cli"
 	"github.com/dbtrail/dbtrail/internal/cliutil"
 	"github.com/dbtrail/dbtrail/internal/config"
 	"github.com/dbtrail/dbtrail/internal/query"
@@ -119,10 +118,9 @@ func init() {
 	reconstructCmd.Flags().StringVar(&recTables, "tables", "", "Comma-separated schema.table list for --output-format=mydumper (e.g. mydb.orders,mydb.users)")
 	reconstructCmd.Flags().StringVar(&recChunkSize, "chunk-size", "256MB", "Max size per SQL chunk file in full-table mode (e.g. 64MB, 1GB)")
 	reconstructCmd.Flags().IntVar(&recParallelism, "parallelism", 0, "Max tables to reconstruct concurrently in full-table mode (default: runtime.NumCPU())")
-	cli.AddDuckDBTuningFlags(reconstructCmd)
-	bindCommandEnv(reconstructCmd)
+	AddDuckDBTuningFlags(reconstructCmd)
+	BindCommandEnv(reconstructCmd)
 
-	rootCmd.AddCommand(reconstructCmd)
 }
 
 func runReconstruct(cmd *cobra.Command, args []string) error {
@@ -269,7 +267,7 @@ func runReconstruct(cmd *cobra.Command, args []string) error {
 		Since:    &snapshotTime,
 		Until:    &at,
 	}
-	duckTuning, err := cli.DuckDBTuningFromFlags(cmd)
+	duckTuning, err := DuckDBTuningFromFlags(cmd)
 	if err != nil {
 		return err
 	}
@@ -278,7 +276,7 @@ func runReconstruct(cmd *cobra.Command, args []string) error {
 		DBName:         dbName,
 		NoArchive:      recNoArchive,
 		AllowGaps:      recAllowGaps,
-		ArchiveFetcher: cli.TunedArchiveFetcher(duckTuning),
+		ArchiveFetcher: TunedArchiveFetcher(duckTuning),
 	})
 	if err != nil {
 		// Surface CLI hints only at the CLI layer; the library types
@@ -450,7 +448,7 @@ func runReconstructFullTable(cmd *cobra.Command, start time.Time) error {
 	}
 
 	// ── Parse --chunk-size ─────────────────────────────────────────────────
-	chunkSize, err := parseByteSize(recChunkSize)
+	chunkSize, err := cliutil.ParseByteSize(recChunkSize)
 	if err != nil {
 		return fmt.Errorf("--chunk-size: %w", err)
 	}
@@ -472,7 +470,7 @@ func runReconstructFullTable(cmd *cobra.Command, start time.Time) error {
 		baselineSrc = recBaselineS3
 	}
 
-	duckTuning, err := cli.DuckDBTuningFromFlags(cmd)
+	duckTuning, err := DuckDBTuningFromFlags(cmd)
 	if err != nil {
 		return err
 	}
@@ -487,7 +485,7 @@ func runReconstructFullTable(cmd *cobra.Command, start time.Time) error {
 		ChunkSize:      chunkSize,
 		Parallelism:    recParallelism,
 		AllowGaps:      recAllowGaps,
-		ArchiveFetcher: cli.TunedArchiveFetcher(duckTuning),
+		ArchiveFetcher: TunedArchiveFetcher(duckTuning),
 	}
 	reports, err := reconstruct.ReconstructTables(cmd.Context(), cfg)
 	if err != nil {
