+++ internal/cli/reconstruct_pk_types_integration_test.go
@@ -1,6 +1,6 @@
 //go:build integration
 
-package main
+package cli
 
 import (
 	"context"
